Changes after rebasing to master.
The changes are mostly in spark3. They are necessitated by the catalog
support introduced in https://github.com/apache/iceberg/pull/1783.
As the spark3 IcebergSource now implements SupportsCatalogOptions,
DataFrameReader#load no longer calls IcebergSource#getTable but calls
SparkCatalog#loadTable directly. In order for the SparkTable returned by
SparkCatalog#loadTable(Identifier) to be aware of the snapshot, the
information about the snapshot needs to be present in the Identifier.
For this reason, we introduce a SnapshotAwareIdentifier interface
extending Identifier.
As SupportsCatalogOptions does not allow a schema to be specified
(requested), SparkTable no longer needs a requestedSchema field, so
some dead code is removed from it.