Address feedback from Yufei Gu.
Undo changes to BitmapPositionDeleteIndex and Deletes.toPositionIndex methods.
There is also no need for DeleteCounter.increment(long).
Add incrementDeleteCount method to DeleteFilter to increment its DeleteCounter
from objects that hold a DeleteFilter and do not have direct access to a
DeleteCounter. Use this in ColumnnarBatchReader to increment the DeleteCounter
when building the row id mapping from positional deletes and applying
equality deletes to it. (Also perform the logic of applying equality deletes
in ColumnarBatchReader.ColumnBatchLoader rather than in DeleteFilter.)